requirements.txt
pyrogram>=2.0.0
tgcrypto>=1.2.0
python-dotenv>=0.19.0
yt-dlp>=2023.3.4
pytgcalls>=0.9.7
mysql-connector-python>=8.0.26
psutil>=5.8.0